backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
pydantic>=2.6,<3.0
python-multipart>=0.0.9
Pillow>=10.0,<12.0
torch>=2.2,<3.0
torchvision>=0.17,<1.0
numpy>=1.26,<3.0
